backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.27
alembic==1.13.1
pymysql==1.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
bcrypt==4.1.2
cryptography==42.0.2
telethon==1.34.0
redis==5.0.1
aioredis==2.0.1
fastapi-cache2==0.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
aiogram==3.3.0
python-dotenv==1.0.1
websockets==12.0
pydantic>=2.4.1,<2.6.0
python-telegram-bot==20.8
aiohttp==3.9.3
pytest==8.0.0
httpx==0.26.0 